requirements.txt
# Core ML/Scientific Computing
# PyTorch with CUDA 12.1 support (compatible with CUDA 12.3 driver)
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy==1.26.4  # Pinned for PyTorch compatibility (NumPy 2.x causes crashes)
scipy>=1.7.0

# Image Processing
Pillow>=9.0.0
rawpy>=0.18.0
opencv-python>=4.8.0  # For advanced image processing
scikit-image>=0.21.0  # For image metrics and analysis
scikit-learn>=1.0.0  # For preprocessing and machine learning utilities
pyiqa>=0.1.7  # For NIQE and other no-reference quality metrics
lpips>=0.1.4  # For perceptual similarity metrics
torchmetrics>=1.0.0  # For PyTorch-native SSIM, PSNR, MSE metrics

# Astronomy/Scientific Data
astropy>=5.0.0
h5py>=3.8.0  # For HDF5 data handling

# Configuration and Utilities
PyYAML>=6.0
tqdm>=4.64.0
psutil>=5.9.0
omegaconf>=2.3.0  # For configuration management
lz4>=4.3.0  # For fast compression
backports.lzma>=0.0.14  # For lzma compression support

# Logging and Monitoring
tensorboard>=2.10.0
wandb>=0.13.0

# Development and Testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
isort>=5.10.0
flake8>=5.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Optional: For advanced optimizations (commented out as they can be fragile)
# triton>=2.0.0  # For custom CUDA kernels
# flash-attn>=2.0.0  # For memory-efficient attention

# Development utilities
ipython>=8.0.0
jupyter>=1.0.0

# HPC and Multi-GPU Training
# Note: torchrun is built into PyTorch - no additional package needed
